Needed to Change Int to Double

+++ src/simulator/control/StateComparator.java
@@ -6,7 +6,7 @@ import org.json.JSONObject;
 public interface StateComparator {
 	boolean equal(JSONObject s1, JSONObject s2);
 
-	static boolean tEqual(JSONObject s1, JSONObject s2){ return s1.getInt("time") == s2.getInt("time"); }
+	static boolean tEqual(JSONObject s1, JSONObject s2){ return s1.getDouble("time") == s2.getDouble("time"); }
 
 	static boolean idEqual(int i, JSONArray ja1, JSONArray ja2){ return ja1.getJSONObject(i).get("id").equals(ja2.getJSONObject(i).get("id")); }
